% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
Main server function for the Shiny application. This is where all data is
loaded and all modules are called. It serves as the central hub connecting
the UI with the various analysis modules.
}
