% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{boxplot_server}
\alias{boxplot_server}
\title{Box Plot Module - Server}
\usage{
boxplot_server(
  id,
  data,
  value_var = "AVAL",
  param_var = "PARAM",
  visit_var = "AVISIT",
  trt_var = "TRTA",
  title = "Box Plot"
)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{A reactive data frame containing the analysis dataset}

\item{value_var}{Name of the value column (default: "AVAL")}

\item{param_var}{Name of the parameter column (default: "PARAM")}

\item{visit_var}{Name of the visit column (default: "AVISIT")}

\item{trt_var}{Name of the treatment column (default: "TRTA")}

\item{title}{Plot title (default: "Box Plot")}
}
\description{
Server logic for the Box Plot module. Filters data based on selected
parameter, visits, and treatments, then renders a faceted box plot using
ggplot2.
}
