% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{boxplot_ui}
\alias{boxplot_ui}
\title{Box Plot Module - UI}
\usage{
boxplot_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A bslib card containing the box plot interface
}
\description{
UI for the Box Plot module. Displays a box plot of vital signs or lab
parameters by treatment and visit. Includes a screenshot button for easy
export of the plot as PNG.
}
