% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cox.R
\name{cox_server}
\alias{cox_server}
\title{Cox PH Model Module - Server}
\usage{
cox_server(id, data, time_var = "AVAL", censor_var = "CNSR", trt_var = "TRTA")
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{A reactive data frame containing the survival analysis dataset}

\item{time_var}{Name of the time-to-event variable (default: "AVAL")}

\item{censor_var}{Name of the censoring indicator variable (default: "CNSR")}

\item{trt_var}{Name of the treatment variable (default: "TRTA")}
}
\description{
Server logic for the Cox Proportional Hazards module. Fits a Cox PH model
using the survival package and renders a forest plot using survminer.
}
