% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demographics.R
\name{demographics_server}
\alias{demographics_server}
\title{Baseline Demographics Module - Server}
\usage{
demographics_server(id, data)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{A reactive data frame containing the demographics dataset}
}
\description{
Server logic for the Baseline Demographics module. Generates a summary
table using gtsummary and renders it with gt.
}
