% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_histogram.R
\name{histogram_ui}
\alias{histogram_ui}
\title{Histogram Module - UI}
\usage{
histogram_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A \code{bslib::card} containing the histogram controls and plot output
}
\description{
Creates the user interface for the Histogram module. Users can select a
parameter, one or more visits, and treatments to view the distribution of
values. Includes a camera button to download the plot as a PNG image.
}
