% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{home_ui}
\alias{home_ui}
\title{Home Page Module - UI}
\usage{
home_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A Shiny UI element (div) containing the welcome page content
}
\description{
This is the landing/home page of the application. It provides an
overview of the app, explains its purpose, describes the datasets used,
and lists all available modules.
}
