% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_km.R
\name{km_server}
\alias{km_server}
\title{Kaplan Meier Curve Module - Server}
\usage{
km_server(id, data, time_var = "AVAL", censor_var = "CNSR", trt_var = "TRTA")
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{A reactive data frame containing the time-to-event dataset}

\item{time_var}{Name of the time variable (default: "AVAL")}

\item{censor_var}{Name of the censoring indicator (default: "CNSR")}

\item{trt_var}{Name of the treatment variable (default: "TRTA")}
}
\description{
Server logic for the Kaplan Meier module. Fits a survival curve using
\code{survfit()} and renders it with risk table using the ggsurvfit package.
}
