% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_meantrend.R
\name{meantrend_ui}
\alias{meantrend_ui}
\title{Mean Trend Plot Module - UI}
\usage{
meantrend_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A bslib card containing the mean trend plot interface
}
\description{
UI for the Mean Trend module. Displays a line plot showing the mean
value of a parameter over visits by treatment group. Includes a screenshot
button to export the plot as PNG.
}
